some changes

+++ main.py
@@ -1,28 +1,21 @@
 from PyQt5.QtWidgets import (
+    QTableWidgetItem,
+    QStyleFactory,
     QApplication,
     QMainWindow,
-    QStyleFactory,
-    QTableWidget,
-    QTableWidgetItem,
     QMessageBox,
     QHeaderView,
 )
-from PyQt5.QtGui import (
-    QFont,
-    QIcon,
-    QPalette,
-    QColor,
-    QImage,
-    QBrush,
-    QDesktopServices,
-)
-from PyQt5.QtCore import Qt, QSize, QUrl
+from PyQt5.QtGui import QIcon, QDesktopServices
+from PyQt5.QtCore import QSize, QUrl
 from bs4 import BeautifulSoup as bs
 from window import Ui_MainWindow
 import requests
 import shutil
+import time
 import sys
 import os
+
 
 HEADERS = {
     "User-Agent": "Mozilla/5.0 (Windows NT 10.0; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/77.0.3865.120 YaBrowser/19.10.2.195 Yowser/2.5 Safari/537.36"
@@ -30,8 +23,6 @@ HEADERS = {
 TABLE_HEADERS = ["Poster", "Title", "Released", "Popularity", "Rating", "Link"]
 SITE = "https://vndb.org"
 IMGS_PATH = os.getcwd() + "\media\\"
-if not os.path.exists(IMGS_PATH):
-    os.mkdir(IMGS_PATH)
 
 
 class mywindow(QMainWindow):
@@ -39,29 +30,28 @@ class mywindow(QMainWindow):
         super(mywindow, self).__init__()
         self.ui = Ui_MainWindow()
         self.ui.setupUi(self)
-        # self.size_horizontal = 420
-        # self.size_vertical = 200
-        self.iconWidth = 70
-        self.iconHeight = 210
-        self.title = "VNDB (test)"
+        self.iconWidth = 80
+        self.iconHeight = 240
+        self.title = "VNDB parser (for fun)"
         self.initUI()
 
     def initUI(self):
-        # self.setFixedSize(self.size_horizontal, self.size_vertical)
-        self.ui.tableWidget.setIconSize(QSize(self.iconWidth, self.iconHeight))
         self.setWindowTitle(
             f"{self.title} – {QApplication.applicationName()} {QApplication.applicationVersion()}"
         )
+
+        self.ui.tableWidget.setIconSize(QSize(self.iconWidth, self.iconHeight))
         self.ui.tableWidget.setHorizontalHeaderLabels(TABLE_HEADERS)
-        for key, value in zip([0, 1, 2, 3, 4], [80, 500, 100, 80, 80]):
+        for key, value in zip([0, 1, 2, 3, 4], [88, 500, 100, 80, 80]):
             self.ui.tableWidget.setColumnWidth(key, value)
         for i in range(6):
             self.ui.tableWidget.horizontalHeader().setSectionResizeMode(
                 i, QHeaderView.Fixed
             )
-
         self.ui.tableWidget.doubleClicked.connect(self.openVnLink)
-        self.ui.searchButton.clicked.connect(self.checkWarnings)
+
+        self.ui.searchInp.returnPressed.connect(self.ui.searchBtn.click)
+        self.ui.searchBtn.clicked.connect(self.checkWarnings)
         self.ui.exitButton.clicked.connect(self.exit)
 
     def openVnLink(self):
@@ -73,17 +63,21 @@ class mywindow(QMainWindow):
             QDesktopServices.openUrl(QUrl(linkVn))
 
     def checkWarnings(self):
-        query = self.ui.searchInput.text()
+        query = self.ui.searchInp.text()
         if query == "":
             QMessageBox.warning(self, "Error", "Заполните ввод!")
             return 0
+
         resultsList = self.parseSearchResults(query)
         if resultsList == False:
             QMessageBox.information(self, "No result", "Ничего не найдено")
             return 0
+
         self.addListItems(resultsList)
+        return 1
 
     def downloadImg(self, link):
+
         html = requests.get(link, headers=HEADERS)
         soup = bs(html.content, "html.parser")
         try:
@@ -93,14 +87,19 @@ class mywindow(QMainWindow):
             imgPath = IMGS_PATH + imgName
             if os.path.exists(imgPath):
                 return imgPath
+
             with open(imgPath, "wb") as f:
                 f.write(requests.get(imgLink, headers=HEADERS).content)
             return imgPath
+
         except AttributeError as e:
             print("LOG:", e)
-            return os.getcwd() + "errorImage.jpg"
+            return os.getcwd() + "\\errorImage.jpg"
 
     def addListItems(self, resultsList):
+
+        if not os.path.exists(IMGS_PATH):
+            os.mkdir(IMGS_PATH)
         for row, item in enumerate(resultsList, start=0):
             titleLink = SITE + item.find("a").get("href")
             imgPath = QIcon(self.downloadImg(titleLink))
@@ -122,16 +121,18 @@ class mywindow(QMainWindow):
                 self.ui.tableWidget.setItem(
                     row, j, QTableWidgetItem(titleParameters[j - 1])
                 )
-        # self.ui.tableWidget.resizeColumnsToContents()
 
     def parseSearchResults(self, query):
+
         html = requests.get(SITE + f"/v?q={query}", headers=HEADERS)
         soup = bs(html.content, "html.parser")
         if soup.find("table") is None:
-            return False
+            return self.parseOnePage()
+
         pagesNum = soup.find("a", text="last »")
         if pagesNum is None:
             return [tr for tr in soup.find_all("tr")][1:]
+
         pages = pagesNum.get("href").split("&")[1]
         pagesCount = int(pages[2:])
         resultsList = []
@@ -142,13 +143,16 @@ class mywindow(QMainWindow):
             soup = bs(html.content, "html.parser")
             for item in soup.find_all("tr")[1:]:
                 resultsList.append(item)
+            time.sleep(1)
         return resultsList
 
     def parseOnePage(self):
+
         return 0
 
     def exit(self):
-        shutil.rmtree(IMGS_PATH)
+        if os.path.exists(IMGS_PATH):
+            shutil.rmtree(IMGS_PATH)
         self.close()
